Reorganize AIP METS generation into own module
